% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{as_conlen}
\alias{as_conlen}
\alias{as_conlen.NULL}
\alias{as_conlen.Seqinfo}
\alias{as_conlen.character}
\alias{as_conlen.data.frame}
\title{Convert to a contig length (Seqinfo) object}
\usage{
as_conlen(x, ...)

\method{as_conlen}{NULL}(x, ...)

\method{as_conlen}{Seqinfo}(x, ...)

\method{as_conlen}{character}(x, ...)

\method{as_conlen}{data.frame}(x, ...)
}
\arguments{
\item{x}{Thing to be converted}

\item{...}{Additional arguments}
}
\description{
Convert to a contig length (Seqinfo) object
}
