% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{as_gff}
\alias{as_gff}
\alias{as_gff.NULL}
\alias{as_gff.GFF}
\alias{as_gff.GRanges}
\alias{as_gff.character}
\alias{as_gff.data.frame}
\title{Convert to a synder GFF object}
\usage{
as_gff(x, ...)

\method{as_gff}{NULL}(x, ...)

\method{as_gff}{GFF}(x, ...)

\method{as_gff}{GRanges}(x, seqinfo_ = NULL, source = NULL, type = NULL,
  score = NULL, phase = NULL, id = NULL, ...)

\method{as_gff}{character}(x, ...)

\method{as_gff}{data.frame}(x, seqinfo_ = NULL, ...)
}
\arguments{
\item{x}{Thing to be converted}

\item{...}{Additional arguments}

\item{seqinfo_}{Seqinfo object}

\item{source}{character vector for GFF column source meta-column}

\item{type}{character vector for GFF column type meta-column}

\item{score}{numeric vector for GFF column score meta-column}

\item{phase}{integer vector for GFF column phase meta-column}

\item{id}{character vector for GFF column attr, is currently used as the
name of the query}
}
\description{
as_gff.character will attempt to load a file, which is expected to be
TAB-delimited and headerless.
}
