% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{as_synmap}
\alias{as_synmap}
\alias{as_synmap.Synmap}
\alias{as_synmap.character}
\alias{as_synmap.Axt}
\alias{as_synmap.data.frame}
\alias{as_synmap.GRangePairs}
\title{Convert to a synder Synmap object}
\usage{
as_synmap(x, ...)

\method{as_synmap}{Synmap}(x, ...)

\method{as_synmap}{character}(x, ...)

\method{as_synmap}{Axt}(x, seqinfo_a = NULL, seqinfo_b = NULL, ...)

\method{as_synmap}{data.frame}(x, seqinfo_a = NULL, seqinfo_b = NULL, ...)

\method{as_synmap}{GRangePairs}(x, seqinfo_a = NULL, seqinfo_b = NULL, ...)
}
\arguments{
\item{x}{Thing to be converted}

\item{...}{Additional arguments}

\item{seqinfo_a}{Seqinfo object for first GRanges entry}

\item{seqinfo_b}{Seqinfo object for second GRanges entry}
}
\description{
Convert to a synder Synmap object
}
