% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_count}
\alias{c_count}
\title{count overlaps}
\usage{
c_count(synfilename, gfffilename)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{gfffilename}{gff file name}
}
\description{
count overlaps
}

