% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_dump}
\alias{c_dump}
\title{print all blocks with contiguous set ids}
\usage{
c_dump(syn, swap, trans, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{trans}{score transform methods, single character}

\item{offsets}{6-element integer vector of [01] offsets}
}
\description{
print all blocks with contiguous set ids
}

