% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_map}
\alias{c_map}
\title{trace intervals across genomes}
\usage{
c_map(syn, gff, swap, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{gff}{gff file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}
}
\description{
trace intervals across genomes
}

