% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_search}
\alias{c_search}
\title{predict search intervals}
\usage{
c_search(synfilename, gfffilename, tclfilename, qclfilename, swap, k, r, trans,
  offsets)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{gfffilename}{GFF file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{k}{STUB}

\item{r}{STUB}

\item{trans}{stuB}
}
\description{
predict search intervals
}

