% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count}
\alias{count}
\title{count overlaps}
\usage{
count(synfilename, intfilename)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
count overlaps
}

