% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{dump}
\alias{dump}
\title{print all blocks with contiguous set ids}
\usage{
dump(synfile, swap = FALSE, trans = "i")
}
\arguments{
\item{filename}{synteny map file name}
}
\description{
print all blocks with contiguous set ids
}

