% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{featureMap}
\alias{featureMap}
\title{Map features between GFFs based on a synder SearchResult}
\usage{
featureMap(srcres, tgff, fgff)
}
\arguments{
\item{srcres}{SearchResult object from a synder::search call}

\item{tgff}{GFF object for target species}

\item{fgff}{GFF object for focal species}
}
\description{
Map features between GFFs based on a synder SearchResult
}
