% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter}
\alias{filter}
\title{remove links that disagree with the synteny map}
\usage{
filter(synfilename, intfilename)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
remove links that disagree with the synteny map
}

