% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{find_obfuscations}
\alias{find_obfuscations}
\title{Find queries that are in syntenically scambled locations}
\usage{
find_obfuscations(x)
}
\arguments{
\item{x}{SearchResult object}
}
\description{
Find queries that are in syntenically scambled locations
}
