% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{find_truncations}
\alias{find_truncations}
\title{Find queries that map to the ends of a scaffold}
\usage{
find_truncations(x)
}
\arguments{
\item{x}{SearchResult object}
}
\description{
Find queries that map to the ends of a scaffold
}
