% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{flag_summary}
\alias{flag_summary}
\title{Summarize the flags returned from a search}
\usage{
flag_summary(x)
}
\arguments{
\item{x}{SearchResult object}
}
\description{
Summarize the flags returned from a search
}
