% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intro_1}
\alias{intro_1}
\title{Intro vignette dataset 1}
\format{List}
\usage{
intro_1
}
\description{
Contains 2 files:
1. synmap - simple linear synteny map
2. tgen   - lengths of target scaffold
}
\keyword{datasets}

