% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intro_2}
\alias{intro_2}
\title{Intro vignette dataset 2}
\format{List}
\usage{
intro_2
}
\description{
Similar to intro_1, but with an inversion
}
\details{
Contains 2 files:
\enumerate{
    \item synmap - simple linear synteny map
    \item tgen   - lengths of target scaffold
 }
}
\keyword{datasets}
