% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{is_incoherent}
\alias{is_incoherent}
\title{Determine which search intervals are not in well-behaved syntenic locations}
\usage{
is_incoherent(x)
}
\arguments{
\item{x}{SearchResult object}
}
\value{
logical vector
}
\description{
Determine which search intervals are not in well-behaved syntenic locations
}
