% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{is_unassembled}
\alias{is_unassembled}
\title{Determine which queries map off the ends of a scaffold}
\usage{
is_unassembled(x)
}
\arguments{
\item{x}{SearchResult object}
}
\description{
Determine which queries map off the ends of a scaffold
}
