% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{load_blastp_file}
\alias{load_blastp_file}
\title{Load a protein BLAST file}
\usage{
load_blastp_file(blastp_file, fmap = NULL, tmap = NULL)
}
\arguments{
\item{blastp_file}{filename}

\item{fmap}{data.frame: a map from blastp qseqid (a protein ID) to the seqid
in the SearchResult (probably a gene ID for an interval)}

\item{tmap}{data.frame: like fmap but for the target proteins}
}
\value{
data.frame
}
\description{
Load a protein BLAST file
}
