% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{load_tblastn_file}
\alias{load_tblastn_file}
\title{Load a tBLASTn file}
\usage{
load_tblastn_file(tblastn_file, fmap = NULL)
}
\arguments{
\item{tblastn_file}{filename}

\item{fmap}{an optional map from target protein to gene name (fseqid)}
}
\description{
Load a tBLASTn file
}
