% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_blastp_map}
\alias{make_blastp_map}
\title{Map intervals for protein blast}
\usage{
make_blastp_map(blastp_file, fgff, tgff, srcres, fmap = NULL,
  tmap = NULL, ffromAttr = identity, tfromAttr = identity)
}
\arguments{
\item{blastp_file}{filename: with columns {qseqid, sseqid, qlen, slen, length, evalue, ppos}}

\item{fgff}{GFF: focal species}

\item{tgff}{GFF: target species}

\item{srcres}{SearchResult object (from \code{.tidy_searchResult})}

\item{fmap}{data.frame: a map from blastp qseqid (a protein ID) to the seqid
in the SearchResult (probably a gene ID for an interval)}

\item{tmap}{data.frame: like fmap but for the target proteins}

\item{ffromAttr}{function (optional) extract IDs attr column in focal GFF}

\item{tfromAttr}{function (optional) extract IDs attr column in target GFF}
}
\value{
list containing the blast results filtered through synder and the raw blast results
}
\description{
Returns info on the focal and target intervals where 1) there is a
significant protein BLAST hit to the proteins encoded by the focal and
target features and 2) the hit overlaps a search interval on the target
side. The columns \code{fseqid} and \code{si_fseqid} hold the protein
encoding focal feature and the focal interval used in the synder search,
respectively. These two will not always be the same.
}
