% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_tblastn_gene_map}
\alias{make_tblastn_gene_map}
\title{make gene map}
\usage{
make_tblastn_gene_map(tblastn_file, tgff, fmap = NULL)
}
\arguments{
\item{tblastn_file}{filename for tblastn results}

\item{tgff}{a tidy GFF file (tidied with .tidy_gff)}

\item{fromAttr}{a function for extracting the sequence ID from the GFF attribute column}
}
\description{
make gene map
}
