% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_tblastn_si_map}
\alias{make_tblastn_si_map}
\title{tBLASTn search interval map}
\usage{
make_tblastn_si_map(tblastn, synres)
}
\arguments{
\item{tblastn}{data.frame}

\item{synres}{data.frame}
}
\description{
tBLASTn search interval map
}
