% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{map}
\alias{map}
\title{trace intervals across genomes}
\usage{
map(synfilename, gfffilename, swap = defaults$swap,
  offsets = defaults$offsets)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
trace intervals across genomes
}

