% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{neighborhood_madness}
\alias{neighborhood_madness}
\title{Calculate the local scrambledness in the region around each query}
\usage{
neighborhood_madness(x, synmap)
}
\arguments{
\item{x}{SearchResult object}

\item{synmap}{Synmap object}
}
\description{
Calculate the local scrambledness in the region around each query
}
