% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DumpResult}
\alias{plot.DumpResult}
\title{Print the result of dumping a synteny map}
\usage{
\method{plot}{DumpResult}(x, ...)
}
\arguments{
\item{x}{DumpResult object}

\item{...}{Additional arguments (not currently used)}
}
\description{
Print the result of dumping a synteny map
}
