% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SearchResult}
\alias{plot.SearchResult}
\title{Plot the result of a search operations}
\usage{
\method{plot}{SearchResult}(x, y, rank = FALSE, ...)
}
\arguments{
\item{x}{SearchResult object}

\item{y}{Synmap object}

\item{rank}{logical Should intervals be printed by rank?}

\item{...}{Additional arguments (not currently used)}
}
\description{
This plot function takes both a SearchResult object and the Synmap object
that was used to produce it. The latter is required to plot the search
intervals in the synmap context.
}
