% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Synmap}
\alias{plot.Synmap}
\title{Plot a synteny map}
\usage{
\method{plot}{Synmap}(x, ...)
}
\arguments{
\item{x}{Synmap object}

\item{...}{Additional arguments (not currently used)}
}
\description{
Plot a synteny map
}
