% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_hitmap}
\alias{read_hitmap}
\title{Read a hit file}
\usage{
read_hitmap(file)
}
\arguments{
\item{file}{hit file name}
}
\value{
A dataframe
}
\description{
A hit file is required to have the same first 6 columns as a synteny map,
i.e. qcon, qstart, qstop, tcon, tstart, tstop. There can be any number of
additional columns.
}
