% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search}
\alias{search}
\title{predict search intervals}
\usage{
search(synfilename, intfilename)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{gfffilename}{GFF file name}
}
\description{
predict search intervals
}

