% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\docType{package}
\name{synder}
\alias{synder}
\alias{synder-package}
\title{synder: Trace intervals between genomes}
\description{
A synteny map is a set of linked intervals between two genomes. The primary
function of Synder is to use a synteny map to trace an interval in one
genome to a narrow search space on another genome.
}
\details{
The primary function exported by Synder are

\itemize{
  \item search - map intervals in A to search intervals in B
  \item anon_search - simplified version of search
  \item dump - dump synteny map with added contiguous set ids
}
}
