% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{synder_plot}
\alias{synder_plot}
\alias{plot.synmap}
\alias{plot.gff}
\alias{plot.hitmap}
\alias{plot.dump_result}
\alias{plot.search_result}
\alias{plot.filter_result}
\alias{plot.map_result}
\alias{plot.count_result}
\title{Synder plot functions}
\usage{
plot.synmap(x, ...)

plot.gff(x, ...)

plot.hitmap(x, ...)

plot.dump_result(x, ...)

plot.search_result(x, y, ...)

plot.filter_result(x, y, ...)

plot.map_result(x, y, ...)

plot.count_result(x, y, ...)
}
\arguments{
\item{x}{table of a synder class}

\item{...}{additional arguments that are currently ignored}

\item{y}{synmap object (for context)}
}
\description{
plots synder classes
}
