% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{synder_print}
\alias{print.count_result}
\alias{print.dump_result}
\alias{print.filter_result}
\alias{print.gff}
\alias{print.hitmap}
\alias{print.map_result}
\alias{print.search_result}
\alias{print.synmap}
\alias{synder_print}
\title{Synder print functions}
\usage{
\method{print}{synmap}(x)

\method{print}{gff}(x)

\method{print}{hitmap}(x)

\method{print}{dump_result}(x)

\method{print}{search_result}(x)

\method{print}{filter_result}(x)

\method{print}{map_result}(x)

\method{print}{count_result}(x)
}
\arguments{
\item{x}{table of a synder class}
}
\description{
Prints the special attributes of the class and then prints a parent class.
}

