% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synder_toy}
\alias{synder_toy}
\title{Toy dataset for synder}
\format{List}
\usage{
synder_toy
}
\description{
Contains 4 files:
1. synmap     - hand made synteny map
2. query_gff  - set of intervals relative to the query
3. target_gff - set of intervals relative to the target
4. hitmap     - map to filter through the main synteny map
}
\keyword{datasets}

