% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{syntenic_density}
\alias{syntenic_density}
\title{Calculate a metric for the overall density of the synteny map}
\usage{
syntenic_density(x, qgen = NULL, tget = NULL)
}
\arguments{
\item{x}{Synmap object}

\item{qgen}{query-side scaffold lengths}

\item{tgen}{target-side scaffold lengths}
}
\description{
Calculate a metric for the overall density of the synteny map
}
