% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{syntenic_scatter}
\alias{syntenic_scatter}
\title{Calculate a metric for the overall scrambledness of the synteny map}
\usage{
syntenic_scatter(x, qgen = NULL, tget = NULL)
}
\arguments{
\item{x}{Synmap object}

\item{qgen}{query-side scaffold lengths}

\item{tgen}{target-side scaffold lengths}
}
\description{
Calculate a metric for the overall scrambledness of the synteny map
}
