% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy}
\alias{toy}
\title{Toy dataset for synder}
\format{List}
\usage{
toy
}
\description{
Contains 4 files:
1. synmap - hand made synteny map
2. qgff   - set of intervals relative to the query
3. tgff   - set of intervals relative to the target
4. hitmap - map to filter through the main synteny map
5. qgen   - lengths of query contigs/scaffolds/chromosomes
t. tgen   - lengths of target contigs/scaffolds/chromosomes
}
\keyword{datasets}

