% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ty1}
\alias{ty1}
\title{Yeast Ty1 mobile genetic element dataset}
\format{List}
\usage{
ty1
}
\description{
Contains 2 files:
\enumerate{
    \item synmaps - Maps between S. cerivisiae and 7 other yeasts
    \item gff     - Positions of Ty1 mobile elements in S. cerivisiae
}
}
\keyword{datasets}
