% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{anon_search}
\alias{anon_search}
\title{Wrapper for search allowing simple interval inputs}
\usage{
anon_search(syn, a, b, seqid, ...)
}
\arguments{
\item{syn}{synteny map file or object}

\item{a, b}{start and stop locations}

\item{seqid}{the name of the reference query contig}

\item{...}{additional arguments sent to synder::search}
}
\description{
The function \code{search} allows powerful mapping to target search
intervals using a feature table (GFF). This approach requires building a GFF
table, which is a needless hassle if you want to search your own intervals
(rather than known features). \code{anon_search} addresses this problem by
allowing searches given only start and stop positions and contig name.
}
