% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_count}
\alias{c_count}
\title{count overlaps}
\usage{
c_count(synfilename, gfffilename, swap, offsets)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{gfffilename}{gff file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}
}
\description{
count overlaps
}

