% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_count}
\alias{c_count}
\title{count overlaps}
\usage{
c_count(syn, gff, swap, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{gff}{GFF file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{offsets}{6-element integer vector of [01] offsets}
}
\description{
count overlaps
}

