% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_filter}
\alias{c_filter}
\title{remove links that disagree with the synteny map}
\usage{
c_filter(synfilename, intfilename, swap, k, r, trans)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{k}{STUB}

\item{r}{STUB}

\item{trans}{stuB}
}
\description{
remove links that disagree with the synteny map
}

