% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_filter}
\alias{c_filter}
\title{remove links that disagree with the synteny map}
\usage{
c_filter(syn, hit, swap, k, r, trans, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{hit}{int file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{k}{match fuziness, integer}

\item{r}{score decay rate, 0 means no context, high means more context}

\item{trans}{score transform methods, single character}

\item{offsets}{4-element integer vector of [01] offsets (start/stop
offsets for the synteny maps and the GFF)}
}
\description{
remove links that disagree with the synteny map
}
