% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_map}
\alias{c_map}
\title{trace intervals across genomes}
\usage{
c_map(syn, gff, swap, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{gff}{GFF file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{offsets}{4-element integer vector of [01] offsets (start/stop
offsets for the synteny maps and the GFF)}
}
\description{
trace intervals across genomes
}
