% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_search}
\alias{c_search}
\title{predict search intervals}
\usage{
c_search(synfilename, gfffilename)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{gfffilename}{GFF file name}
}
\description{
predict search intervals
}

