% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_search}
\alias{c_search}
\title{predict search intervals}
\usage{
c_search(syn, gff, tcl, qcl, swap, k, r, trans, offsets)
}
\arguments{
\item{syn}{synteny map file name}

\item{gff}{GFF file name}

\item{tcl}{target contig lengths file name}

\item{qcl}{target contig lengths file name}

\item{swap}{reverse direction of synteny map (e.g. swap query and target)}

\item{k}{match fuziness, integer}

\item{r}{score decay rate, 0 means no context, high means more context}

\item{trans}{score transform methods, single character}

\item{offsets}{6-element integer vector of [01] offsets}
}
\description{
predict search intervals
}
