% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{count}
\alias{count}
\title{count overlaps}
\usage{
count(synfilename, gfffilename, swap = FALSE)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
count overlaps
}

