% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{featureMap}
\alias{featureMap}
\title{Map features between GFFs based on a synder SearchResult}
\usage{
featureMap(srcres, tgff, fgff, tfromAttr = identity,
  ffromAttr = identity)
}
\arguments{
\item{srcres}{tidy search result (from .tidy_searchResult)}

\item{tgff}{data.frame for GFF object for target species (from .tidy_gff)}

\item{fgff}{data.frame for GFF object for focal species (from .tidy_gff)}

\item{tfromAttr}{function (optional) extract IDs attr column in target GFF}

\item{ffromAttr}{function (optional) extract IDs attr column in focal GFF}
}
\description{
Map features between GFFs based on a synder SearchResult
}
