% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{filter}
\alias{filter}
\title{remove links that disagree with the synteny map}
\usage{
filter(synfilename, intfilename, swap = defaults$swap,
  trans = defaults$trans, k = defaults$k, r = defaults$r,
  offsets = defaults$offsets)
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
remove links that disagree with the synteny map
}

