% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intro_1}
\alias{intro_1}
\title{Intro vignette dataset 1}
\format{List}
\usage{
intro_1
}
\description{
Contains 2 files:
\enumerate{
    \item synmap - simple linear synteny map with seqinfo for both sides
 }
}
\keyword{datasets}
