% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_tblastn_gene_map}
\alias{make_tblastn_gene_map}
\title{make gene map}
\usage{
make_tblastn_gene_map(tblastn, tgff)
}
\arguments{
\item{tblastn}{data.frame}

\item{tgff}{data.frame}
}
\description{
make gene map
}
