% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_tblastn_gene_map}
\alias{make_tblastn_gene_map}
\title{make gene map}
\usage{
make_tblastn_gene_map(tblastn_file, tgff, fmap = NULL,
  fromAttr = identity)
}
\arguments{
\item{tblastn_file}{filename for tblastn results}

\item{tgff}{GFF file or object}

\item{fmap}{an optional map from target protein to gene name (fseqid)}

\item{fromAttr}{function mapping 'attr' column to gene name}
}
\value{
data.frame
}
\description{
make gene map
}
