% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relate.R
\name{make_tblastn_si_map}
\alias{make_tblastn_si_map}
\title{tBLASTn search interval map}
\usage{
make_tblastn_si_map(tblastn_file, srcres, fmap = NULL)
}
\arguments{
\item{tblastn_file}{filename of tblastn result}

\item{srcres}{data.frame}

\item{fmap}{an optional map from target protein to gene name (fseqid)}
}
\description{
tBLASTn search interval map
}
