% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsynder.R
\name{search}
\alias{search}
\title{trace intervals to target search intervals}
\usage{
search(synfilename, gfffilename, tclfile = "", qclfile = "", swap = FALSE,
  k = 0, r = 0, trans = "i")
}
\arguments{
\item{synfilename}{synteny map file name}

\item{intfilename}{int file name}
}
\description{
trace intervals to target search intervals
}

