% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{synder_cast}
\alias{synder_cast}
\alias{as.data.frame.Synmap}
\alias{as.data.frame.DumpResult}
\alias{as.data.frame.SearchResult}
\alias{as.data.frame.Seqinfo}
\alias{as.data.frame.GFF}
\title{Synder cast functions}
\usage{
\method{as.data.frame}{Synmap}(x, ...)

\method{as.data.frame}{DumpResult}(x, ...)

\method{as.data.frame}{SearchResult}(x, ...)

\method{as.data.frame}{Seqinfo}(x, ...)

\method{as.data.frame}{GFF}(x, ...)
}
\arguments{
\item{...}{additional arguments}

\item{d}{input type}
}
\description{
Cast data as synder objects
}
\details{
\code{as_synmap}
\code{as_gff}
\code{as_conlen}
}
