% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{synder_cast}
\alias{synder_cast}
\alias{as_synmap}
\alias{as_gff}
\alias{as_conlen}
\title{Synder cast functions}
\usage{
as_synmap(x, ...)

as_gff(x, ...)

as_conlen(x, ...)
}
\arguments{
\item{...}{additional arguments}

\item{d}{input type}
}
\description{
Cast data as synder objects
}
\details{
\code{as_synmap}
\code{as_gff}
\code{as_conlen}
}
