% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{synder_plot}
\alias{synder_plot}
\alias{plot.synmap}
\alias{plot.gff}
\alias{plot.hitmap}
\alias{plot.dump_result}
\alias{plot.search_result}
\alias{plot.filter_result}
\alias{plot.map_result}
\alias{plot.count_result}
\title{Synder plot functions}
\usage{
\method{plot}{synmap}(x, ...)

\method{plot}{gff}(x, ...)

\method{plot}{hitmap}(x, ...)

\method{plot}{dump_result}(x, ...)

\method{plot}{search_result}(x, y, ...)

\method{plot}{filter_result}(x, y, ...)

\method{plot}{map_result}(x, y, ...)

\method{plot}{count_result}(x, y, ...)
}
\arguments{
\item{x}{table of a synder class}

\item{...}{additional arguments that are currently ignored}

\item{y}{synmap object (for context)}
}
\description{
plots synder classes
}
