% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{synder_print}
\alias{synder_print}
\alias{print.Synmap}
\alias{print.GFF}
\alias{print.DumpResult}
\alias{print.SearchResult}
\title{Synder print functions}
\usage{
\method{print}{Synmap}(x, ...)

\method{print}{GFF}(x, ...)

\method{print}{DumpResult}(x, ...)

\method{print}{SearchResult}(x, ...)
}
\arguments{
\item{x}{table of a synder class}

\item{...}{additional arguments that are currently ignored}
}
\description{
Prints the special attributes of the class and then prints a parent class.
}
