% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{synder_read}
\alias{synder_read}
\alias{read_synmap}
\alias{read_gff}
\alias{read_conlen}
\title{Synder read functions}
\usage{
read_synmap(file, ...)

read_gff(file, ...)

read_conlen(file, ...)
}
\arguments{
\item{file}{TAB-delimited file}

\item{...}{Additional arguments for the casting function (\code{as_*})}
}
\description{
read synder classes from files
}
