% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy}
\alias{toy}
\title{Toy dataset for synder}
\format{List}
\usage{
toy
}
\description{
Contains 4 files:
\enumerate{
    \item synmap - hand made synteny map with seqinfo for both sides
    \item qgff   - set of intervals relative to the query
    \item tgff   - set of intervals relative to the target
}
}
\keyword{datasets}
